LayoutTests/platform/mac/accessibility/static-text-role-uses-text-under-element-expected.txt

Text 1 Text 2 Text 3
This tests that if an element uses the 'text' role but does not use an aria-label, it can retrieve the text inside of it.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS text.stringValue is 'AXValue:   Text 3'
PASS successfullyParsed is true

TEST COMPLETE
